check_deposit_utxo: Move it from utils to extended_rpc.
